% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_heat_pulse_velocity_core.R
\name{add_quality_flags}
\alias{add_quality_flags}
\title{Add Quality Flags to Results}
\usage{
add_quality_flags(results)
}
\arguments{
\item{results}{Data frame with velocity results}
}
\value{
Data frame with added quality_flag column
}
\description{
Add Quality Flags to Results
}
\keyword{internal}
