% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_heat_pulse_velocity_advanced.R
\name{advanced_hpv_methods}
\alias{advanced_hpv_methods}
\title{Advanced Heat Pulse Velocity Methods}
\description{
Implementation of Compensation Heat Pulse Method (CHPM) and
Double Ratio Method (DRM) for sap flow analysis
}
