% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reporting.R
\name{apply_export_filters}
\alias{apply_export_filters}
\title{Apply Export Filters}
\usage{
apply_export_filters(vh_results, filter_options)
}
\arguments{
\item{vh_results}{A vh_results object}

\item{filter_options}{List of filtering options}
}
\value{
Filtered vh_results object
}
\description{
Internal function to apply filtering options to velocity results
}
\keyword{internal}
