% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reporting.R
\name{apply_export_filters_sap_data}
\alias{apply_export_filters_sap_data}
\title{Apply Export Filters to Sap Data}
\usage{
apply_export_filters_sap_data(sap_data, filter_options)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{filter_options}{List of filtering options}
}
\value{
Filtered sap_data object
}
\description{
Internal function to apply filtering options to sap data
}
\keyword{internal}
