% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_heat_pulse_velocity_advanced.R
\name{assess_temp_data_quality}
\alias{assess_temp_data_quality}
\title{Assess Temperature Data Quality for HPV Methods}
\usage{
assess_temp_data_quality(temp_data, min_signal_ratio = 3)
}
\arguments{
\item{temp_data}{temperature time series data}

\item{min_signal_ratio}{minimum signal-to-noise ratio (default 3)}
}
\value{
list with quality metrics
}
\description{
Evaluate the quality of temperature data for heat pulse calculations.
}
