% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04g_sap_flux_plots.R
\name{auto_label}
\alias{auto_label}
\title{Auto-generate axis label from column name}
\usage{
auto_label(col_name)
}
\arguments{
\item{col_name}{Character. Column name}
}
\value{
Character. Formatted label
}
\description{
Auto-generate axis label from column name
}
\keyword{internal}
