% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_heat_pulse_velocity_core.R
\name{calc_dma}
\alias{calc_dma}
\title{Calculate DMA velocities}
\usage{
calc_dma(hrm_results, tmax_klu_results, diffusivity, probe_spacing)
}
\description{
Calculate DMA velocities
}
\keyword{internal}
