% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_heat_pulse_velocity_yaml.R
\name{calc_heat_pulse_velocity_yaml}
\alias{calc_heat_pulse_velocity_yaml}
\title{Calculate Heat Pulse Velocity with YAML Methods}
\usage{
calc_heat_pulse_velocity_yaml(
  sap_data,
  methods,
  parameters = NULL,
  yaml_methods_dir = NULL
)
}
\arguments{
\item{sap_data}{A sap_data object from read_sap_data()}

\item{methods}{Character vector of method names to use}

\item{parameters}{List of calculation parameters}

\item{yaml_methods_dir}{Directory containing YAML method files}
}
\value{
A tibble containing calculated heat pulse velocities
}
\description{
Main function to calculate heat pulse velocity using YAML-defined methods.
}
