% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_heat_pulse_velocity_core.R
\name{calc_hrm}
\alias{calc_hrm}
\title{Calculate HRM velocities}
\usage{
calc_hrm(
  dTratio_douo,
  dTratio_diui,
  HRM_period,
  diffusivity,
  probe_spacing,
  tp
)
}
\description{
Calculate HRM velocities
}
\keyword{internal}
