% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_heat_pulse_velocity_advanced.R
\name{calc_method_compatibility}
\alias{calc_method_compatibility}
\title{Calculate Method Compatibility Matrix}
\usage{
calc_method_compatibility(probe_configs)
}
\arguments{
\item{probe_configs}{list of probe configuration objects}
}
\value{
data.frame with method compatibility
}
\description{
Generate a matrix showing which methods are compatible with
different probe configurations and conditions.
}
