% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_heat_pulse_velocity_core.R
\name{calc_mhr}
\alias{calc_mhr}
\title{Calculate MHR velocities}
\usage{
calc_mhr(
  deltaT_do,
  deltaT_di,
  deltaT_uo,
  deltaT_ui,
  diffusivity,
  probe_spacing
)
}
\description{
Calculate MHR velocities
}
\keyword{internal}
