% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_heat_pulse_velocity_core.R
\name{calc_tmax_coh}
\alias{calc_tmax_coh}
\title{Calculate Tmax Cohen velocities}
\usage{
calc_tmax_coh(
  deltaT_do,
  deltaT_di,
  diffusivity,
  probe_spacing,
  pre_pulse,
  peak_info
)
}
\arguments{
\item{peak_info}{Pre-computed peak information from calc_vh_single_pulse}
}
\description{
Calculate Tmax Cohen velocities
}
\keyword{internal}
