% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_heat_pulse_velocity_yaml.R
\name{calc_vh_single_pulse_yaml}
\alias{calc_vh_single_pulse_yaml}
\title{Calculate velocity for single pulse using YAML methods}
\usage{
calc_vh_single_pulse_yaml(
  pulse_id,
  measurements,
  parameters,
  methods,
  yaml_methods_dir = NULL
)
}
\arguments{
\item{pulse_id}{ID of pulse to process}

\item{measurements}{Data frame with measurement data}

\item{parameters}{List of parameters}

\item{methods}{Character vector of method names}

\item{yaml_methods_dir}{Directory containing YAML method files}
}
\value{
Data frame with results for this pulse
}
\description{
Calculate velocity for single pulse using YAML methods
}
\keyword{internal}
