% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reporting.R
\name{calculate_method_correlations}
\alias{calculate_method_correlations}
\title{Calculate Method Correlations}
\usage{
calculate_method_correlations(data_df)
}
\arguments{
\item{data_df}{Data frame with velocity results}
}
\value{
Correlation matrix
}
\description{
Internal function to calculate correlation matrix between methods
}
\keyword{internal}
