% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03f_thermal_diffusivity.R
\name{calculate_time_to_max_temperatures}
\alias{calculate_time_to_max_temperatures}
\title{Calculate Time to Maximum Temperature}
\usage{
calculate_time_to_max_temperatures(zero_temps, pre_pulse = 30)
}
\arguments{
\item{zero_temps}{Data frame of temperature measurements}

\item{pre_pulse}{Pre-pulse period (sec)}
}
\value{
List with:
\item{t_max_downstream}{Vector of times to max for downstream (outer: do, inner: di)}
\item{t_max_upstream}{Vector of times to max for upstream (outer: uo, inner: ui)}
}
\description{
Finds time to maximum temperature for downstream and upstream probes
during zero-flow periods.
}
\keyword{internal}
