% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_data_validation.R
\name{calculate_validation_summary}
\alias{calculate_validation_summary}
\title{Calculate Validation Summary Statistics}
\usage{
calculate_validation_summary(heat_pulse_data, issues, warnings)
}
\arguments{
\item{heat_pulse_data}{A heat_pulse_data object}

\item{issues}{Character vector of issues}

\item{warnings}{Character vector of warnings}
}
\value{
List with summary statistics
}
\description{
Calculate Validation Summary Statistics
}
\keyword{internal}
