% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{check_package_status}
\alias{check_package_status}
\title{Check Package Version and Dependencies}
\usage{
check_package_status()
}
\value{
A list with version information and dependency status
}
\description{
Utility function to check package version and dependency status
}
\examples{
\dontrun{
check_package_status()
}

}
