% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_tree_utilities.R
\name{cohens_d}
\alias{cohens_d}
\title{Calculate Cohen's d effect size}
\usage{
cohens_d(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
Cohen's d value
}
\description{
Calculate Cohen's d effect size
}
\keyword{internal}
