% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reporting.R
\name{combine_data_for_csv}
\alias{combine_data_for_csv}
\title{Combine Data for CSV Export}
\usage{
combine_data_for_csv(export_data)
}
\arguments{
\item{export_data}{List with export data}
}
\value{
Data frame with combined data
}
\description{
Internal function to combine multiple data tables for export
}
\keyword{internal}
