% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{combine_multiple_heat_pulse_data}
\alias{combine_multiple_heat_pulse_data}
\title{Combine multiple heat_pulse_data objects into unified structure}
\usage{
combine_multiple_heat_pulse_data(
  individual_data,
  import_summary,
  show_progress
)
}
\arguments{
\item{individual_data}{List of heat_pulse_data objects}

\item{import_summary}{Data frame with import summary}

\item{show_progress}{Logical, whether to show progress}
}
\value{
Combined heat_pulse_data object
}
\description{
Combine multiple heat_pulse_data objects into unified structure
}
\keyword{internal}
