% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{combine_multiple_sap_data}
\alias{combine_multiple_sap_data}
\title{Combine multiple sap_data objects into unified structure}
\usage{
combine_multiple_sap_data(individual_data, import_summary, show_progress)
}
\arguments{
\item{individual_data}{List of sap_data objects}

\item{import_summary}{Data frame with import summary}

\item{show_progress}{Logical, whether to show progress}
}
\value{
Combined sap_data object
}
\description{
Combine multiple sap_data objects into unified structure
}
\keyword{internal}
