% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_tree_utilities.R
\name{compare_trees_temporal}
\alias{compare_trees_temporal}
\title{Generate temporal comparison of trees}
\usage{
compare_trees_temporal(velocity_results, sap_data, time_window)
}
\arguments{
\item{velocity_results}{Data frame with velocity results}

\item{sap_data}{Multi-tree sap_data object}

\item{time_window}{Character, time window for aggregation}
}
\value{
List with temporal analysis results
}
\description{
Generate temporal comparison of trees
}
\keyword{internal}
