% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{convert_markdown_to_html}
\alias{convert_markdown_to_html}
\title{Convert Markdown to HTML}
\usage{
convert_markdown_to_html(markdown_content)
}
\arguments{
\item{markdown_content}{Character vector with markdown content}
}
\value{
Character vector with HTML content
}
\description{
Internal function for basic markdown to HTML conversion
}
\keyword{internal}
