% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{create_empty_diagnostics}
\alias{create_empty_diagnostics}
\title{Create empty diagnostics data frame}
\usage{
create_empty_diagnostics()
}
\value{
Empty data frame with correct structure
}
\description{
Create empty diagnostics data frame
}
\keyword{internal}
