% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_import.R
\name{create_empty_measurements}
\alias{create_empty_measurements}
\title{Create empty measurements data frame}
\usage{
create_empty_measurements()
}
\value{
Empty data frame with correct structure
}
\description{
Create empty measurements data frame
}
\keyword{internal}
