% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{create_empty_sap_data}
\alias{create_empty_sap_data}
\title{Create empty sap_data object for failed imports}
\usage{
create_empty_sap_data(tree_id, file_path)
}
\arguments{
\item{tree_id}{Character, tree identifier}

\item{file_path}{Character, file path}
}
\value{
Empty sap_data object
}
\description{
Create empty sap_data object for failed imports
}
\keyword{internal}
