% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{create_histogram_plot}
\alias{create_histogram_plot}
\title{Create Histogram Plot}
\usage{
create_histogram_plot(vh_results, file_path)
}
\arguments{
\item{vh_results}{A vh_results object}

\item{file_path}{Output file path}
}
\description{
Internal function to create histogram plot
}
\keyword{internal}
