% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a_method_calibration_enhanced.R
\name{create_quadratic_plot}
\alias{create_quadratic_plot}
\title{Create Quadratic Model Plot}
\usage{
create_quadratic_plot(quad_result, primary_method, secondary_method)
}
\arguments{
\item{quad_result}{Quadratic model result object}

\item{primary_method}{Primary method name}

\item{secondary_method}{Secondary method name}
}
\value{
ggplot object
}
\description{
Creates diagnostic plot for piecewise linear-quadratic model
}
\keyword{internal}
