% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{create_quality_plot}
\alias{create_quality_plot}
\title{Create Quality Plot}
\usage{
create_quality_plot(vh_results, file_path)
}
\arguments{
\item{vh_results}{A vh_results object}

\item{file_path}{Output file path}
}
\description{
Internal function to create quality flags plot
}
\keyword{internal}
