% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01k_dual_stability_gradient_correction.R
\name{create_segments_from_dates}
\alias{create_segments_from_dates}
\title{Create Segments from Dates}
\usage{
create_segments_from_dates(changepoint_dates, date_range)
}
\arguments{
\item{changepoint_dates}{Vector of changepoint dates}

\item{date_range}{Range of all dates (min, max)}
}
\value{
Data frame with segment information
}
\description{
Creates segment data frame from changepoint dates.
}
\keyword{internal}
