% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_heat_pulse_velocity_core.R
\name{create_test_sap_data}
\alias{create_test_sap_data}
\title{Create Test Sap Data Object}
\usage{
create_test_sap_data(n_points = 120, n_pulses = 1, add_noise = TRUE)
}
\arguments{
\item{n_points}{Number of temperature measurements per pulse}

\item{n_pulses}{Number of pulses to simulate}

\item{add_noise}{Whether to add random noise to temperature readings}
}
\value{
A sap_data object compatible with package functions
}
\description{
Creates a properly formatted sap_data object for testing purposes.
}
\keyword{internal}
