% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{create_timeseries_plot}
\alias{create_timeseries_plot}
\title{Create Time Series Plot}
\usage{
create_timeseries_plot(vh_results, file_path)
}
\arguments{
\item{vh_results}{A vh_results object}

\item{file_path}{Output file path}
}
\description{
Internal function to create time series plot
}
\keyword{internal}
