% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01l_vpd_changepoints.R
\name{create_vpd_segments}
\alias{create_vpd_segments}
\title{Create VPD Segments}
\usage{
create_vpd_segments(changepoint_dates, date_range)
}
\arguments{
\item{changepoint_dates}{Vector of changepoint dates}

\item{date_range}{Range of all dates (min, max)}
}
\value{
Data frame with segment information
}
\description{
Creates segment data frame from changepoint dates.
}
\keyword{internal}
