% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04f_daily_aggregation.R
\name{detect_interval}
\alias{detect_interval}
\title{Detect measurement interval from datetime stamps}
\usage{
detect_interval(datetime)
}
\arguments{
\item{datetime}{POSIXct vector}
}
\value{
Numeric. Interval in hours
}
\description{
Detect measurement interval from datetime stamps
}
\keyword{internal}
