% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_sapwood_area.R
\name{estimate_bark_thickness}
\alias{estimate_bark_thickness}
\title{Estimate Bark Thickness from DBH and Species}
\usage{
estimate_bark_thickness(dbh, species_group, method)
}
\arguments{
\item{dbh}{Diameter at breast height (cm)}

\item{species_group}{Species group}

\item{method}{Estimation method}
}
\value{
Estimated bark thickness (cm)
}
\description{
Estimate Bark Thickness from DBH and Species
}
\keyword{internal}
