% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_sapwood_area.R
\name{estimate_heartwood_radius}
\alias{estimate_heartwood_radius}
\title{Estimate Heartwood Radius from Cambium Radius and Species}
\usage{
estimate_heartwood_radius(cambium_radius, species_group, method)
}
\arguments{
\item{cambium_radius}{Cambium radius (cm)}

\item{species_group}{Species group}

\item{method}{Estimation method}
}
\value{
Estimated heartwood radius (cm)
}
\description{
Estimate Heartwood Radius from Cambium Radius and Species
}
\keyword{internal}
