% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02a_data_filtering.R
\name{estimate_pulse_interval}
\alias{estimate_pulse_interval}
\title{Estimate Pulse Interval}
\usage{
estimate_pulse_interval(datetimes)
}
\arguments{
\item{datetimes}{POSIXct vector of timestamps}
}
\value{
Numeric, estimated interval in hours
}
\description{
Estimates typical time between pulses from datetime vector.
}
\keyword{internal}
