% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_heat_pulse_velocity_yaml.R
\name{execute_calculation_steps}
\alias{execute_calculation_steps}
\title{Execute Calculation Steps}
\usage{
execute_calculation_steps(steps, env)
}
\arguments{
\item{steps}{List of calculation steps}

\item{env}{Environment containing input variables}
}
\value{
List containing calculation results
}
\description{
Executes the calculation steps defined in the method configuration.
}
\keyword{internal}
