% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01l_vpd_changepoints.R
\name{filter_by_minimum_spacing}
\alias{filter_by_minimum_spacing}
\title{Filter Indices by Minimum Spacing}
\usage{
filter_by_minimum_spacing(indices, dates, vpd_values, min_spacing_days)
}
\arguments{
\item{indices}{Vector of candidate indices}

\item{dates}{Vector of dates corresponding to indices}

\item{vpd_values}{VPD values corresponding to indices}

\item{min_spacing_days}{Minimum days between selected indices}
}
\value{
Filtered vector of indices
}
\description{
Ensures minimum spacing between selected indices, keeping those with lowest VPD.
}
\keyword{internal}
