% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01g_ui_helpers.R
\name{format_parameter_summary}
\alias{format_parameter_summary}
\title{Format Parameter Summary for Display}
\usage{
format_parameter_summary(
  probe_config,
  wood_properties,
  params,
  methods,
  n_pulses,
  k_source = NULL
)
}
\arguments{
\item{probe_config}{ProbeConfiguration object}

\item{wood_properties}{WoodProperties object}

\item{params}{List of calculation parameters}

\item{methods}{Character vector of methods}

\item{n_pulses}{Number of pulses to process}
}
\value{
Character string with formatted summary
}
\description{
Format Parameter Summary for Display
}
\keyword{internal}
