% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_data_overview}
\alias{generate_data_overview}
\title{Generate Data Overview}
\usage{
generate_data_overview(sap_data, vh_results)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{vh_results}{A vh_results object}
}
\value{
Character vector with overview text
}
\description{
Internal function to generate data overview section
}
\keyword{internal}
