% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_executive_summary}
\alias{generate_executive_summary}
\title{Generate Executive Summary}
\usage{
generate_executive_summary(sap_data, vh_results)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{vh_results}{A vh_results object}
}
\value{
Character vector with summary text
}
\description{
Internal function to generate executive summary section
}
\keyword{internal}
