% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_quality_assessment}
\alias{generate_quality_assessment}
\title{Generate Quality Assessment}
\usage{
generate_quality_assessment(vh_results)
}
\arguments{
\item{vh_results}{A vh_results object}
}
\value{
Character vector with quality assessment text
}
\description{
Internal function to generate quality assessment section
}
\keyword{internal}
