% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_raw_data_tables}
\alias{generate_raw_data_tables}
\title{Generate Raw Data Tables}
\usage{
generate_raw_data_tables(sap_data, vh_results)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{vh_results}{A vh_results object}
}
\value{
Character vector with raw data tables text
}
\description{
Internal function to generate raw data tables section
}
\keyword{internal}
