% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_recommendations}
\alias{generate_recommendations}
\title{Generate Recommendations}
\usage{
generate_recommendations(sap_data, vh_results)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{vh_results}{A vh_results object}
}
\value{
Character vector with recommendations text
}
\description{
Internal function to generate recommendations section
}
\keyword{internal}
