% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_report_plots}
\alias{generate_report_plots}
\title{Generate Report Plots}
\usage{
generate_report_plots(vh_results, figures_dir)
}
\arguments{
\item{vh_results}{A vh_results object}

\item{figures_dir}{Directory to save figures}
}
\value{
List with figure file paths
}
\description{
Internal function to generate all plots for report
}
\keyword{internal}
