% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_report_tables}
\alias{generate_report_tables}
\title{Generate Report Tables}
\usage{
generate_report_tables(sap_data, vh_results, report_type)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{vh_results}{A vh_results object}

\item{report_type}{Character string}
}
\value{
List with table data
}
\description{
Internal function to generate all tables for report
}
\keyword{internal}
