% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_technical_details}
\alias{generate_technical_details}
\title{Generate Technical Details}
\usage{
generate_technical_details(sap_data, vh_results)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{vh_results}{A vh_results object}
}
\value{
Character vector with technical details text
}
\description{
Internal function to generate technical details section
}
\keyword{internal}
