% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/09_diagnostic_reporting.R
\name{generate_temporal_analysis}
\alias{generate_temporal_analysis}
\title{Generate Temporal Analysis}
\usage{
generate_temporal_analysis(vh_results)
}
\arguments{
\item{vh_results}{A vh_results object}
}
\value{
Character vector with temporal analysis text
}
\description{
Internal function to generate temporal analysis section
}
\keyword{internal}
