% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_data_import.R
\name{generate_tree_ids}
\alias{generate_tree_ids}
\title{Generate tree IDs from file paths}
\usage{
generate_tree_ids(file_paths)
}
\arguments{
\item{file_paths}{Character vector of file paths}
}
\value{
Character vector of tree IDs
}
\description{
Generate tree IDs from file paths
}
\keyword{internal}
