% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03h_flux_density.R
\name{get_sap_flux_conversion_factor}
\alias{get_sap_flux_conversion_factor}
\title{Get Sap Flux Conversion Factor from Wood Properties}
\usage{
get_sap_flux_conversion_factor(wood_properties)
}
\arguments{
\item{wood_properties}{WoodProperties R6 object}
}
\value{
Numeric Z factor value
}
\description{
Extracts the Z factor from a WoodProperties object for manual calculations.
}
\examples{
\dontrun{
wood_props <- load_wood_properties("eucalyptus")
wood_props <- calculate_wood_properties(wood_props)

Z <- get_sap_flux_conversion_factor(wood_props)
print(Z)
}

}
\seealso{
Other flux density functions: 
\code{\link{apply_flux_conversion}()},
\code{\link{calc_becker_edwards_k}()},
\code{\link{calc_sap_flux_density}()},
\code{\link{sap_flux_conversion_factor_at_temp}()}
}
\concept{flux density functions}
\keyword{internal}
