% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_temporal_processing.R
\name{interpolate_group_data}
\alias{interpolate_group_data}
\title{Interpolate data for a single group}
\usage{
interpolate_group_data(group_data, max_gap_hours)
}
\arguments{
\item{group_data}{Data frame for a single method/sensor combination}

\item{max_gap_hours}{Maximum gap size to interpolate}
}
\value{
Data frame with interpolated values
}
\description{
Interpolate data for a single group
}
\keyword{internal}
