% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reporting.R
\name{interpret_normality_tests}
\alias{interpret_normality_tests}
\title{Interpret Normality Tests}
\usage{
interpret_normality_tests(shapiro_result, ks_result)
}
\arguments{
\item{shapiro_result}{Shapiro test result}

\item{ks_result}{Kolmogorov-Smirnov test result}
}
\value{
Character string with interpretation
}
\description{
Internal function to interpret normality test results
}
\keyword{internal}
