% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{is_shiny}
\alias{is_shiny}
\title{Detect if running in Shiny environment}
\usage{
is_shiny()
}
\value{
Logical indicating if in Shiny
}
\description{
Detect if running in Shiny environment
}
\keyword{internal}
