% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_heat_pulse_velocity_advanced.R
\name{optimize_drm_window}
\alias{optimize_drm_window}
\title{Optimise Time Window for DRM}
\usage{
optimize_drm_window(
  temp_data,
  probe_spacing,
  thermal_diffusivity = 0.0025,
  sensor_noise = 0.02,
  window_sizes = seq(5, 100, 5)
)
}
\arguments{
\item{temp_data}{data.frame with temperature time series}

\item{probe_spacing}{probe configuration}

\item{thermal_diffusivity}{thermal diffusivity estimate}

\item{sensor_noise}{sensor noise level}

\item{window_sizes}{vector of window sizes to test (default 5 to 100 seconds)}
}
\value{
list with optimal window parameters
}
\description{
Determine the optimal time window for DRM calculations based on
uncertainty minimization principle from Deng et al. (2021).
}
\details{
From Deng et al. (2021), Equation 10:
SE_DRM = (1/n_t) * sqrt(sum(σ²_DRM(t)))

This finds the time window that minimises the standard error of the mean
estimated velocity based on theoretical uncertainty.
}
