% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04f_daily_aggregation.R
\name{parse_interval}
\alias{parse_interval}
\title{Parse interval string to hours}
\usage{
parse_interval(interval)
}
\arguments{
\item{interval}{Character string like "30 min", "1 hour", etc.}
}
\value{
Numeric. Interval in hours
}
\description{
Parse interval string to hours
}
\keyword{internal}
