% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reporting.R
\name{perform_distribution_tests}
\alias{perform_distribution_tests}
\title{Perform Distribution Tests}
\usage{
perform_distribution_tests(data_df)
}
\arguments{
\item{data_df}{Data frame with velocity results}
}
\value{
List with distribution test results
}
\description{
Internal function for distribution normality tests
}
\keyword{internal}
