% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_heat_pulse_trace}
\alias{plot_heat_pulse_trace}
\title{Plot Heat Pulse Trace with Calculation Windows}
\usage{
plot_heat_pulse_trace(
  heat_pulse_data,
  vh_results,
  pulse_id,
  show_methods = NULL,
  pre_pulse = 30
)
}
\arguments{
\item{heat_pulse_data}{A heat_pulse_data object from read_heat_pulse_data()}

\item{vh_results}{Results tibble from calc_heat_pulse_velocity()}

\item{pulse_id}{Pulse ID to plot}

\item{show_methods}{Character vector of methods to show windows for.
If NULL, shows all methods present in results. Default: NULL}

\item{pre_pulse}{Number of seconds of pre-pulse data to show. Default: 30}
}
\value{
A ggplot2 object
}
\description{
Creates a diagnostic plot showing temperature traces for a single heat pulse
with calculation windows and timepoints overlaid for different methods.
}
\examples{
\dontrun{
heat_pulse_data <- read_heat_pulse_data("data.txt")
results <- calc_heat_pulse_velocity(heat_pulse_data)
plot_heat_pulse_trace(heat_pulse_data, results, pulse_id = 1)

# Show only specific methods
plot_heat_pulse_trace(heat_pulse_data, results, pulse_id = 1,
                      show_methods = c("HRM", "Tmax_Coh"))
}

}
