% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_heat_pulse_velocity_core.R
\name{plot_pulse_temps}
\alias{plot_pulse_temps}
\title{Plot Temperature Differences for Diagnostic Purposes}
\usage{
plot_pulse_temps(
  pulse_data,
  delatT_do,
  delatT_di,
  delatT_uo,
  delatT_ui,
  pulse_id
)
}
\arguments{
\item{pulse_data}{Data for single pulse}

\item{delatT_do}{Delta temperatures for downstream outer}

\item{delatT_di}{Delta temperatures for downstream inner}

\item{delatT_uo}{Delta temperatures for upstream outer}

\item{delatT_ui}{Delta temperatures for upstream inner}

\item{pulse_id}{Pulse ID for title}
}
\description{
Plot Temperature Differences for Diagnostic Purposes
}
\keyword{internal}
