% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01k_dual_stability_gradient_correction.R
\name{print.dual_stable_periods}
\alias{print.dual_stable_periods}
\title{Print Method for Dual Stable Periods}
\usage{
\method{print}{dual_stable_periods}(x, ...)
}
\arguments{
\item{x}{A dual_stable_periods object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Dual Stable Periods
}
