% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_spacing_correction.R
\name{print.heartwood_reference_check}
\alias{print.heartwood_reference_check}
\title{Print Method for Heartwood Reference Check}
\usage{
\method{print}{heartwood_reference_check}(x, ...)
}
\arguments{
\item{x}{A heartwood_reference_check object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Heartwood Reference Check
}
