% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_spacing_correction.R
\name{print.heartwood_reference_correction_result}
\alias{print.heartwood_reference_correction_result}
\title{Print Method for Heartwood Reference Correction Result}
\usage{
\method{print}{heartwood_reference_correction_result}(x, ...)
}
\arguments{
\item{x}{A heartwood_reference_correction_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Heartwood Reference Correction Result
}
