% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02a_data_filtering.R
\name{print.interpolation_preview}
\alias{print.interpolation_preview}
\title{Print Method for Interpolation Preview}
\usage{
\method{print}{interpolation_preview}(x, ...)
}
\arguments{
\item{x}{An interpolation_preview object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Interpolation Preview
}
