% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_thermal_diffusivity.R
\name{print.k_estimation_result}
\alias{print.k_estimation_result}
\title{Print Method for k Estimation Results}
\usage{
\method{print}{k_estimation_result}(x, ...)
}
\arguments{
\item{x}{A k_estimation_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for k Estimation Results
}
