% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a_method_calibration.R
\name{print.method_calibration}
\alias{print.method_calibration}
\title{Print Method for Method Calibration Results}
\usage{
\method{print}{method_calibration}(x, ...)
}
\arguments{
\item{x}{A method_calibration object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Method Calibration Results
}
