% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03e_changepoint_detection.R
\name{print.pelt_changepoints}
\alias{print.pelt_changepoints}
\title{Print PELT Changepoint Result}
\usage{
\method{print}{pelt_changepoints}(x, ...)
}
\description{
Print PELT Changepoint Result
}
\keyword{internal}
