% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/12_sapwood_area.R
\name{print.sapwood_area}
\alias{print.sapwood_area}
\title{Print method for sapwood_area objects}
\usage{
\method{print}{sapwood_area}(x, ...)
}
\arguments{
\item{x}{A sapwood_area object}

\item{...}{Additional arguments}
}
\description{
Print method for sapwood_area objects
}
