% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a_method_calibration.R
\name{print.segmented_comparison}
\alias{print.segmented_comparison}
\title{Print Method for Segmented Comparison Results}
\usage{
\method{print}{segmented_comparison}(x, ...)
}
\arguments{
\item{x}{A segmented_comparison object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Segmented Comparison Results
}
