% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01i_weather_data.R
\name{print.silo_data}
\alias{print.silo_data}
\title{Print method for silo_data objects}
\usage{
\method{print}{silo_data}(x, ...)
}
\arguments{
\item{x}{A silo_data object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for silo_data objects
}
