% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_spacing_correction.R
\name{print.spacing_correction_result}
\alias{print.spacing_correction_result}
\title{Print Method for Spacing Correction Results}
\usage{
\method{print}{spacing_correction_result}(x, ...)
}
\arguments{
\item{x}{A spacing_correction_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Spacing Correction Results
}
