% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01l_vpd_changepoints.R
\name{print.stable_vpd_periods}
\alias{print.stable_vpd_periods}
\title{Print Method for Stable VPD Periods}
\usage{
\method{print}{stable_vpd_periods}(x, ...)
}
\arguments{
\item{x}{A stable_vpd_periods object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for Stable VPD Periods
}
