% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04a_method_calibration.R
\name{print.threshold_optimization_result}
\alias{print.threshold_optimization_result}
\title{Print Method for Threshold Optimization Results}
\usage{
\method{print}{threshold_optimization_result}(x, ...)
}
\arguments{
\item{x}{A threshold_optimization_result object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Threshold Optimization Results
}
