% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_tree_utilities.R
\name{print.tree_comparison}
\alias{print.tree_comparison}
\title{Print tree comparison results}
\usage{
\method{print}{tree_comparison}(x, ...)
}
\arguments{
\item{x}{Tree comparison object}

\item{...}{Additional arguments}
}
\description{
Print tree comparison results
}
