% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01f_data_quality.R
\name{print.vh_quality_results}
\alias{print.vh_quality_results}
\title{Print Method for Vh Quality Results}
\usage{
\method{print}{vh_quality_results}(x, ...)
}
\arguments{
\item{x}{A vh_quality_results object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Vh Quality Results
}
