% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01l_vpd_changepoints.R
\name{print.vpd_changepoints}
\alias{print.vpd_changepoints}
\title{Print Method for VPD Changepoints}
\usage{
\method{print}{vpd_changepoints}(x, ...)
}
\arguments{
\item{x}{A vpd_changepoints object}

\item{...}{Additional arguments (not used)}
}
\description{
Print Method for VPD Changepoints
}
