% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01i_weather_data.R
\name{print.weather_data}
\alias{print.weather_data}
\title{Print method for weather_data objects}
\usage{
\method{print}{weather_data}(x, ...)
}
\arguments{
\item{x}{A weather_data object}

\item{...}{Additional arguments (not used)}
}
\description{
Print method for weather_data objects
}
