% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_wood_configuration.R
\name{print.wood_properties_derivation}
\alias{print.wood_properties_derivation}
\title{Print Method for Wood Properties Derivation}
\usage{
\method{print}{wood_properties_derivation}(x, ...)
}
\arguments{
\item{x}{A wood_properties_derivation object}

\item{...}{Additional arguments (ignored)}
}
\description{
Print Method for Wood Properties Derivation
}
