% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03g_wound_correction.R
\name{print_wound_correction_summary}
\alias{print_wound_correction_summary}
\title{Print Wound Correction Summary}
\usage{
print_wound_correction_summary(
  vh_data,
  wound_diameter_vector,
  B_vector,
  input_col,
  use_temporal = FALSE
)
}
\arguments{
\item{vh_data}{Corrected data frame}

\item{input_col}{Input column name}

\item{wound_diameter}{Wound diameter used}

\item{B}{Correction factor applied}
}
\description{
Prints summary of wound correction results.
}
\keyword{internal}
