% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_heat_pulse_velocity_core.R
\name{prompt_parameter_confirmation}
\alias{prompt_parameter_confirmation}
\title{Prompt User for Parameter Confirmation}
\usage{
prompt_parameter_confirmation(
  probe_config,
  wood_properties,
  params,
  methods,
  n_pulses
)
}
\arguments{
\item{probe_config}{ProbeConfiguration object}

\item{wood_properties}{WoodProperties object}

\item{params}{List of calculation parameters}

\item{methods}{Character vector of methods}

\item{n_pulses}{Number of pulses}
}
\value{
Logical indicating whether user confirmed
}
\description{
Prompt User for Parameter Confirmation
}
\keyword{internal}
