% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_wound_correction.R
\name{prompt_wound_diameter}
\alias{prompt_wound_diameter}
\title{Prompt User for Wound Diameter}
\usage{
prompt_wound_diameter()
}
\value{
Wound diameter in cm
}
\description{
Interactive prompt to get wound diameter from user.
}
\keyword{internal}
