% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{read_csv_format}
\alias{read_csv_format}
\title{Read CSV format (tab or comma delimited)}
\usage{
read_csv_format(file_path, show_progress, ...)
}
\arguments{
\item{file_path}{Path to data file}

\item{show_progress}{Show progress updates}

\item{...}{Additional arguments}
}
\description{
Read CSV format (tab or comma delimited)
}
\keyword{internal}
