% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{read_ict_legacy}
\alias{read_ict_legacy}
\title{Read ICT Legacy format}
\usage{
read_ict_legacy(file_path, chunk_size, show_progress, ...)
}
\arguments{
\item{file_path}{Path to data file}

\item{chunk_size}{Characters per chunk}

\item{show_progress}{Show progress updates}

\item{...}{Additional arguments}
}
\description{
Read ICT Legacy format
}
\keyword{internal}
