% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/05_heat_pulse_velocity_advanced.R
\name{select_best_result}
\alias{select_best_result}
\title{Select Best Result from Multiple Methods}
\usage{
select_best_result(method_results)
}
\arguments{
\item{method_results}{list of results from different methods}
}
\value{
list with best result selection
}
\description{
Choose the most reliable result when multiple methods are available.
}
\keyword{internal}
