% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01a_data_import.R
\name{show_message}
\alias{show_message}
\title{Display message in console or Shiny UI}
\usage{
show_message(msg, type = "message", duration = NULL)
}
\arguments{
\item{msg}{Character string message to display}

\item{type}{Type of notification for Shiny ("default", "message", "warning", "error")}

\item{duration}{Duration in seconds for Shiny notification (NULL for auto)}
}
\description{
Display message in console or Shiny UI
}
\keyword{internal}
