% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reporting.R
\name{standardise_for_research}
\alias{standardise_for_research}
\title{Standardise Data for Research Format}
\usage{
standardise_for_research(velocity_results)
}
\arguments{
\item{velocity_results}{Data frame with velocity results}
}
\value{
Data frame with standardised format
}
\description{
Internal function to standardise data for research publications
}
\keyword{internal}
