% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_thermal_diffusivity.R
\name{thermal_diffusivity}
\alias{thermal_diffusivity}
\title{Thermal Diffusivity Estimation Functions}
\description{
Functions for estimating thermal diffusivity (k) from time to maximum
temperature (Tmax) during zero-flow periods. Used to validate assumed
k values and detect probe asymmetry.
}
