% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_probe_configuration.R
\name{validate_data_structure}
\alias{validate_data_structure}
\title{Validate Data Structure}
\usage{
validate_data_structure(sap_data)
}
\arguments{
\item{sap_data}{A sap_data object}
}
\value{
List with validation results
}
\description{
Validate Data Structure
}
\keyword{internal}
