% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06_heat_pulse_velocity_yaml.R
\name{validate_method_inputs}
\alias{validate_method_inputs}
\title{Validate Method Inputs}
\usage{
validate_method_inputs(inputs, config)
}
\arguments{
\item{inputs}{List containing input variables}

\item{config}{Method configuration list}
}
\description{
Validates that all required inputs are provided and have correct types.
}
\keyword{internal}
