% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_probe_configuration.R
\name{validate_method_requirements}
\alias{validate_method_requirements}
\title{Validate Method Requirements}
\usage{
validate_method_requirements(method_name, method_params, config)
}
\arguments{
\item{method_name}{Name of calculation method}

\item{method_params}{Required parameters for the method}

\item{config}{ProbeConfiguration object}
}
\value{
List with validation results
}
\description{
Validate Method Requirements
}
\keyword{internal}
