% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_data_validation.R
\name{validate_ranges}
\alias{validate_ranges}
\title{Validate Data Ranges}
\usage{
validate_ranges(sap_data, temperature_range, voltage_range)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{temperature_range}{Range for temperature values}

\item{voltage_range}{Range for voltage values}
}
\value{
List with issues and warnings
}
\description{
Validate Data Ranges
}
\keyword{internal}
