% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01c_probe_configuration.R
\name{validate_sensor_alignment}
\alias{validate_sensor_alignment}
\title{Validate Sensor Alignment}
\usage{
validate_sensor_alignment(sap_data, config)
}
\arguments{
\item{sap_data}{A sap_data object}

\item{config}{ProbeConfiguration object}
}
\value{
List with validation results
}
\description{
Validate Sensor Alignment
}
\keyword{internal}
