% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_data_validation.R
\name{validate_structure}
\alias{validate_structure}
\title{Validate Data Structure}
\usage{
validate_structure(sap_data)
}
\arguments{
\item{sap_data}{A sap_data object}
}
\value{
List with issues and warnings
}
\description{
Validate Data Structure
}
\keyword{internal}
