% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01i_weather_data.R
\name{validate_weather_data}
\alias{validate_weather_data}
\title{Validate weather data}
\usage{
validate_weather_data(data)
}
\arguments{
\item{data}{A tibble with weather data}
}
\value{
The input tibble with validation warnings/errors
}
\description{
Validate weather data
}
\keyword{internal}
