% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01d_wood_configuration.R
\name{WoodProperties}
\alias{WoodProperties}
\title{Wood Properties R6 Class}
\description{
Wood Properties R6 Class

Wood Properties R6 Class
}
\details{
R6 class to store and manage wood thermal and physical properties.
This includes measurements, constants, derived properties, wound correction,
tree measurements, and quality thresholds.
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config_name}}{Name of the configuration}

\item{\code{site_location}}{List with latitude, longitude, and timezone for the site}

\item{\code{wood_measurements}}{List of user-provided measurements (Method 1 or Method 2)}

\item{\code{wood_constants}}{List of physical constants (thermal/sap/cell wall properties)}

\item{\code{wood_property}}{List of basic wood identifiers (species, type, temperature)}

\item{\code{tree_measurements}}{List of tree-specific measurements (dbh, sapwood_thickness, etc.)}

\item{\code{wound_correction}}{List of wound configuration (drill bit, temporal tracking)}

\item{\code{quality_thresholds}}{List of quality control thresholds}

\item{\code{derived_properties}}{List of calculated properties (populated by calculate_wood_properties())}

\item{\code{yaml_source}}{Path to source YAML file (if loaded from file)}

\item{\code{yaml_data}}{Raw YAML data (if loaded from file)}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{thermal_diffusivity}}{Thermal diffusivity in cm^2/s (actual if available, otherwise default)}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-WoodProperties-new}{\code{WoodProperties$new()}}
\item \href{#method-WoodProperties-validate}{\code{WoodProperties$validate()}}
\item \href{#method-WoodProperties-print}{\code{WoodProperties$print()}}
\item \href{#method-WoodProperties-is_australia}{\code{WoodProperties$is_australia()}}
\item \href{#method-WoodProperties-get_dawn_times}{\code{WoodProperties$get_dawn_times()}}
\item \href{#method-WoodProperties-clone}{\code{WoodProperties$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WoodProperties-new"></a>}}
\if{latex}{\out{\hypertarget{method-WoodProperties-new}{}}}
\subsection{Method \code{new()}}{
Initialize a new WoodProperties object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WoodProperties$new(
  config_name = "Custom Wood Properties",
  site_location = NULL,
  wood_measurements = NULL,
  wood_constants = NULL,
  wood_property = NULL,
  tree_measurements = NULL,
  wound_correction = NULL,
  quality_thresholds = NULL,
  derived_properties = NULL,
  yaml_source = NULL,
  yaml_data = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_name}}{Name of the configuration}

\item{\code{site_location}}{List with latitude, longitude, timezone for predawn/SILO calculations}

\item{\code{wood_measurements}}{List of measurements (Method 1 or Method 2)}

\item{\code{wood_constants}}{List of physical constants}

\item{\code{wood_property}}{List of basic wood properties}

\item{\code{tree_measurements}}{List of tree measurements}

\item{\code{wound_correction}}{List of wound configuration}

\item{\code{quality_thresholds}}{List of quality thresholds}

\item{\code{derived_properties}}{List of derived properties}

\item{\code{yaml_source}}{Path to source YAML file}

\item{\code{yaml_data}}{Raw YAML data}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WoodProperties-validate"></a>}}
\if{latex}{\out{\hypertarget{method-WoodProperties-validate}{}}}
\subsection{Method \code{validate()}}{
Validate wood properties
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WoodProperties$validate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WoodProperties-print"></a>}}
\if{latex}{\out{\hypertarget{method-WoodProperties-print}{}}}
\subsection{Method \code{print()}}{
Print wood properties summary
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WoodProperties$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WoodProperties-is_australia"></a>}}
\if{latex}{\out{\hypertarget{method-WoodProperties-is_australia}{}}}
\subsection{Method \code{is_australia()}}{
Check whether the site is located within Australia's bounding box.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WoodProperties$is_australia()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Logical scalar; FALSE if site_location is not set.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WoodProperties-get_dawn_times"></a>}}
\if{latex}{\out{\hypertarget{method-WoodProperties-get_dawn_times}{}}}
\subsection{Method \code{get_dawn_times()}}{
Calculate astronomical dawn times for a vector of dates.

Requires the \pkg{suncalc} package (listed in Suggests).
\code{site_location$latitude} and \code{site_location$longitude} must be set.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WoodProperties$get_dawn_times(dates)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{dates}}{A vector of \code{Date} objects.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A POSIXct vector of dawn times in \code{site_location$timezone} (defaults to UTC).
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-WoodProperties-clone"></a>}}
\if{latex}{\out{\hypertarget{method-WoodProperties-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{WoodProperties$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
