% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01j_vpd_changepoints.R
\name{assign_segment_ids}
\alias{assign_segment_ids}
\title{Assign Segment IDs to Daily VPD Data}
\usage{
assign_segment_ids(daily_vpd, changepoint_dates)
}
\arguments{
\item{daily_vpd}{Data frame with date column}

\item{changepoint_dates}{Vector of changepoint dates}
}
\value{
daily_vpd with added segment_id column
}
\description{
Assigns segment IDs based on changepoint dates.
}
\keyword{internal}
