% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_heat_pulse_velocity_core.R
\name{calc_hrmx}
\alias{calc_hrmx}
\title{Calculate HRMX velocities}
\usage{
calc_hrmx(
  delatT_do,
  delatT_di,
  delatT_uo,
  delatT_ui,
  dTratio_douo,
  dTratio_diui,
  L,
  H,
  diffusivity,
  probe_spacing
)
}
\description{
Calculate HRMX velocities
}
\keyword{internal}
