% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_heat_pulse_velocity_core.R
\name{calc_mhr}
\alias{calc_mhr}
\title{Calculate MHR velocities}
\usage{
calc_mhr(delatT_do, delatT_di, delatT_uo, delatT_ui, diffusivity, x)
}
\description{
Calculate MHR velocities
}
\keyword{internal}
