% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_heat_pulse_velocity_core.R
\name{calc_tmax_coh}
\alias{calc_tmax_coh}
\title{Calculate Tmax Cohen velocities}
\usage{
calc_tmax_coh(delatT_do, delatT_di, diffusivity, x, pre_pulse)
}
\description{
Calculate Tmax Cohen velocities
}
\keyword{internal}
