% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_tmax_coh_cpp}
\alias{calc_tmax_coh_cpp}
\title{Calculate Tmax Cohen - C++ Implementation}
\usage{
calc_tmax_coh_cpp(deltaT_do, deltaT_di, diffusivity, probe_spacing, pre_pulse)
}
\arguments{
\item{deltaT_do}{Numeric vector of delta temperatures (downstream outer)}

\item{deltaT_di}{Numeric vector of delta temperatures (downstream inner)}

\item{diffusivity}{Thermal diffusivity (cm²/s)}

\item{probe_spacing}{Probe spacing (cm)}

\item{pre_pulse}{Pre-pulse period (seconds)}
}
\value{
List containing Tmax_Coh results
}
\description{
Calculate Tmax Cohen - C++ Implementation
}
\keyword{internal}
