% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03a_heat_pulse_velocity_core.R
\name{calc_tmax_klu}
\alias{calc_tmax_klu}
\title{Calculate Tmax Kluitenberg velocities}
\usage{
calc_tmax_klu(
  deltaT_do,
  deltaT_di,
  diffusivity,
  probe_spacing,
  tp_1,
  pre_pulse
)
}
\description{
Calculate Tmax Kluitenberg velocities
}
\keyword{internal}
