% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{calc_tmax_klu_cpp}
\alias{calc_tmax_klu_cpp}
\title{Calculate Tmax Kluitenberg - C++ Implementation}
\usage{
calc_tmax_klu_cpp(
  deltaT_do,
  deltaT_di,
  diffusivity,
  probe_spacing,
  tp_1,
  pre_pulse
)
}
\arguments{
\item{deltaT_do}{Numeric vector of delta temperatures (downstream outer)}

\item{deltaT_di}{Numeric vector of delta temperatures (downstream inner)}

\item{diffusivity}{Thermal diffusivity (cm²/s)}

\item{probe_spacing}{Probe spacing (cm)}

\item{tp_1}{Heat pulse duration (seconds)}

\item{pre_pulse}{Pre-pulse period (seconds)}
}
\value{
List containing Tmax_Klu results
}
\description{
Calculate Tmax Kluitenberg - C++ Implementation
}
\keyword{internal}
