% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_heat_pulse_velocity_core.R
\name{calc_vh_single_pulse}
\alias{calc_vh_single_pulse}
\title{Calculate velocity for single pulse}
\usage{
calc_vh_single_pulse(
  pulse_data,
  pulse_id,
  parameters,
  methods,
  plot_results = FALSE
)
}
\arguments{
\item{pulse_data}{Pre-filtered data frame for this pulse}

\item{pulse_id}{ID of pulse to process}

\item{parameters}{List of parameters}

\item{methods}{Character vector of methods}

\item{plot_results}{Whether to plot results}
}
\value{
Data frame with results for this pulse
}
\description{
Calculate velocity for single pulse
}
\keyword{internal}
