% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01i_weather_data.R
\name{calc_vpd}
\alias{calc_vpd}
\title{Calculate vapor pressure deficit (VPD)}
\usage{
calc_vpd(
  weather_data,
  temp_col = "air_temp_c",
  rh_col = "relative_humidity",
  a = 17.27,
  b = 237.7,
  return_components = FALSE
)
}
\arguments{
\item{weather_data}{A weather_data object from \code{read_weather_data()}, or
a data frame with columns \code{air_temp_c} and \code{relative_humidity}.}

\item{temp_col}{Character string. Name of temperature column if not using
weather_data object. Default is "air_temp_c".}

\item{rh_col}{Character string. Name of relative humidity column if not using
weather_data object. Default is "relative_humidity".}

\item{a}{Numeric. Magnus equation parameter 'a'. Default is 17.27.}

\item{b}{Numeric. Magnus equation parameter 'b' (째C). Default is 237.7.}

\item{return_components}{Logical. If TRUE, returns saturated vapor pressure
and actual vapor pressure in addition to VPD. Default is FALSE.}
}
\value{
If \code{return_components = FALSE}, returns the input data frame with
an additional column \code{vpd_kpa} (vapor pressure deficit in kPa).
If \code{return_components = TRUE}, also adds columns \code{svp_kpa}
(saturated vapor pressure) and \code{avp_kpa} (actual vapor pressure).
}
\description{
Calculates vapor pressure deficit from air temperature and relative humidity
using the Magnus equation. VPD is the difference between the amount of moisture
in the air and how much moisture the air can hold when saturated.
}
\details{
The Magnus equation calculates saturated vapor pressure (SVP) as:
\deqn{SVP = 0.61078 \times \exp\left(\frac{a \times T}{b + T}\right)}

Where T is temperature in 째C, a = 17.27, and b = 237.7째C.

Actual vapor pressure (AVP) is calculated from relative humidity:
\deqn{AVP = \frac{RH}{100} \times SVP}

Vapor pressure deficit is then:
\deqn{VPD = SVP - AVP}

Default parameters (a=17.27, b=237.7) are appropriate for temperatures
between 0-60째C. For sub-zero temperatures, consider adjusting parameters.
}
\examples{
\dontrun{
# Using weather_data object
weather <- read_weather_data("weather_station.csv")
weather_vpd <- calc_vpd(weather)

# Using custom data frame
my_data <- data.frame(
  temp = c(20, 25, 30),
  rh = c(50, 60, 70)
)
my_data_vpd <- calc_vpd(my_data, temp_col = "temp", rh_col = "rh")

# Return all components
weather_full <- calc_vpd(weather, return_components = TRUE)

# Custom Magnus parameters
weather_vpd <- calc_vpd(weather, a = 17.27, b = 237.7)
}

}
\references{
Alduchov, O. A., & Eskridge, R. E. (1996). Improved Magnus form approximation
of saturation vapor pressure. Journal of Applied Meteorology, 35(4), 601-609.
}
\seealso{
Other weather data functions: 
\code{\link{read_weather_data}()}
}
\concept{weather data functions}
