% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{parse_date_input}
\alias{parse_date_input}
\title{Parse Date Input from User}
\usage{
parse_date_input(input, reference_date)
}
\arguments{
\item{input}{Character string from user input}

\item{reference_date}{POSIXct reference date for numeric offsets}
}
\value{
POSIXct datetime
}
\description{
Helper function to parse date input which can be a date string or number of days
}
\keyword{internal}
