% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01e_heat_pulse_velocity_core.R
\name{plot_pulse_temps}
\alias{plot_pulse_temps}
\title{Plot Temperature Differences for Diagnostic Purposes}
\usage{
plot_pulse_temps(
  pulse_data,
  deltaT_do,
  deltaT_di,
  deltaT_uo,
  deltaT_ui,
  pulse_id
)
}
\arguments{
\item{pulse_data}{Data for single pulse}

\item{deltaT_do}{Delta temperatures for downstream outer}

\item{deltaT_di}{Delta temperatures for downstream inner}

\item{deltaT_uo}{Delta temperatures for upstream outer}

\item{deltaT_ui}{Delta temperatures for upstream inner}

\item{pulse_id}{Pulse ID for title}
}
\description{
Plot Temperature Differences for Diagnostic Purposes
}
\keyword{internal}
