% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03b_wound_correction.R
\name{print_wound_correction_summary}
\alias{print_wound_correction_summary}
\title{Print Wound Correction Summary}
\usage{
print_wound_correction_summary(vh_data, wound_diameter, B, input_col)
}
\arguments{
\item{vh_data}{Corrected data frame}

\item{wound_diameter}{Wound diameter used}

\item{B}{Correction factor applied}

\item{input_col}{Input column name}
}
\description{
Prints summary of wound correction results.
}
\keyword{internal}
