% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01b_data_validation.R
\name{validate_temporal_consistency}
\alias{validate_temporal_consistency}
\title{Validate Temporal Consistency}
\usage{
validate_temporal_consistency(sap_data)
}
\arguments{
\item{sap_data}{A sap_data object}
}
\value{
List with issues and warnings
}
\description{
Validate Temporal Consistency
}
\keyword{internal}
